Revert "Merge branch 'merger-2' of https://github.com/Daura951/Everybody-Hates-Elijah into merger-2"
This reverts commit 90650150192793dbac0d5b67cbb7571d23467a70, reversing
changes made to 46dac8c704dc4b5f227ac177564821ace64bbccf.
