--- Assets/Animations/Player/Stunned.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7669dc63e76062741b667d2cce3a2887
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
