--- Assets/Animations/Player/Stunned/Air Spin.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 055166a35e8af3e4391edee67c33a593
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
