--- Assets/Art/Player Spritesheets/Stunned.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 80a5301188d99444d956de6d747f4198
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
