--- Assets/Scripts/Enemy Scripts/EnemyHurtVals.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd19e8041c6e9cd4d8e31f91098c353d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
