--- Assets/Scripts/Enemy Scripts/EnemyStun_Info.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fa42b190b61b33347b3c2632eedf9dda
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
