--- Assets/Scripts/EnvironmentHurtVals.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 87fc5c9fa81ef24429ecb6ec7b334ae3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
