--- Assets/Scripts/EnvironmentalStun_info.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 75ac808ef17b22d47a1745c46f1dd029
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
