+++ Assets/Scripts/Stun_Info.cs
@@ -2,12 +2,21 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-//Abstract class that contains the basics of what a stun is. Enemy and environment stun uses ths as a superclass. Makes things more orerly for designers
-public abstract class Stun_Info : MonoBehaviour
+public class Stun_Info : MonoBehaviour
 {
-	protected float[] DAKTInfo;
+	public string DAKT;
+	private float[] DAKTInfo;
 
-	public abstract void CalculateStunVals();
+	void Start()
+	{
+		DAKTInfo = new float[4];
+
+		string[] Info = DAKT.Split(" ");
+        DAKTInfo[0] = float.Parse(Info[0]); //Damage
+        DAKTInfo[1] = float.Parse(Info[1]); //Angle
+        DAKTInfo[2] = float.Parse(Info[2]); //Knockback
+        DAKTInfo[3] = float.Parse(Info[3]); //Time
+	}
 
 	public float[] GetDAKTInfo()
 	{
