+++ Assets/prefabs/Lava Floor.prefab
@@ -11,7 +11,7 @@ GameObject:
   - component: {fileID: 7417413966450255309}
   - component: {fileID: 7417413966450255310}
   - component: {fileID: 7417413966450255311}
-  - component: {fileID: 6946870977438102462}
+  - component: {fileID: 7417413966450255280}
   m_Layer: 0
   m_Name: Lava Floor
   m_TagString: KnockBack
@@ -112,7 +112,7 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!114 &6946870977438102462
+--- !u!114 &7417413966450255280
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -121,8 +121,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 7417413966450255281}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 75ac808ef17b22d47a1745c46f1dd029, type: 3}
+  m_Script: {fileID: 11500000, guid: 6b31fcb4b7ae4624dbadfad6945ed17f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  DAKT: 
-  attackType: 0
+  DAKT: 0 0 0 0
